fix: extract statistics computation in their own file
