+++ app/src/androidTest/java/com/epfl/drawyourpath/database/MockDatabaseTestPhoto.kt
@@ -3,7 +3,7 @@ package com.epfl.drawyourpath.database
 import android.graphics.Bitmap
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.epfl.drawyourpath.chat.Message
-import com.epfl.utils.drawyourpath.Utils
+import com.epfl.drawyourpath.utils.Utils
 import org.junit.Assert
 import org.junit.Test
 import org.junit.runner.RunWith
