+++ app/src/main/java/com/epfl/drawyourpath/database/MockDatabase.kt
@@ -8,7 +8,7 @@ import com.epfl.drawyourpath.chat.MessageContent
 import com.epfl.drawyourpath.path.Path
 import com.epfl.drawyourpath.path.Run
 import com.epfl.drawyourpath.userProfile.dailygoal.DailyGoal
-import com.epfl.utils.drawyourpath.Utils
+import com.epfl.drawyourpath.utils.Utils
 import com.google.android.gms.maps.model.LatLng
 import java.time.LocalDate
 import java.time.LocalTime
