+++ app/src/main/java/com/epfl/drawyourpath/mainpage/fragments/helperClasses/FriendsViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.epfl.drawyourpath.database.Database
-import com.epfl.utils.drawyourpath.Utils
+import com.epfl.drawyourpath.utils.Utils
 
 class FriendsViewModel(private val userId: String, private val database: Database) : ViewModel() {
