+++ app/src/main/java/com/epfl/drawyourpath/path/RunsAdapter.kt
@@ -9,7 +9,7 @@ import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.epfl.drawyourpath.R
-import com.epfl.utils.drawyourpath.Utils.getStaticMapUrl
+import com.epfl.drawyourpath.utils.Utils.getStaticMapUrl
 import com.google.android.gms.maps.model.LatLng
 
 /**
