+++ app/src/main/java/com/epfl/drawyourpath/pathDrawing/PathDrawingCurrentPerformanceFragment.kt
@@ -6,9 +6,9 @@ import android.widget.TextView
 import androidx.fragment.app.Fragment
 import com.epfl.drawyourpath.R
 import com.epfl.drawyourpath.path.Run
-import com.epfl.utils.drawyourpath.Utils.getStringDistance
-import com.epfl.utils.drawyourpath.Utils.getStringDuration
-import com.epfl.utils.drawyourpath.Utils.getStringSpeed
+import com.epfl.drawyourpath.utils.Utils.getStringDistance
+import com.epfl.drawyourpath.utils.Utils.getStringDuration
+import com.epfl.drawyourpath.utils.Utils.getStringSpeed
 
 /**
  * Fragment used to displayed the performance of a user during a run(time, distance, average speed).
