+++ app/src/main/java/com/epfl/drawyourpath/pathDrawing/PathDrawingDetailPerformanceFragment.kt
@@ -6,10 +6,10 @@ import android.widget.TextView
 import androidx.fragment.app.Fragment
 import com.epfl.drawyourpath.R
 import com.epfl.drawyourpath.path.Run
-import com.epfl.utils.drawyourpath.Utils.getStringDistance
-import com.epfl.utils.drawyourpath.Utils.getStringDuration
-import com.epfl.utils.drawyourpath.Utils.getStringSpeed
-import com.epfl.utils.drawyourpath.Utils.getStringTimeStartEnd
+import com.epfl.drawyourpath.utils.Utils.getStringDistance
+import com.epfl.drawyourpath.utils.Utils.getStringDuration
+import com.epfl.drawyourpath.utils.Utils.getStringSpeed
+import com.epfl.drawyourpath.utils.Utils.getStringTimeStartEnd
 
 /**
  * Fragment used to displayed the deatil performance of a user during a run(time, distance, average speed, start time, end time, time for 1 km, calories).
